Poc_files/requirements.txt
aiofiles==24.1.0
anyio==4.6.2.post1
asn1crypto==1.5.1
attrs==24.2.0
beautifulsoup4==4.12.3
boto3==1.35.47
botocore==1.35.47
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
cryptography==43.0.3
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.16.1
fitz==0.0.1.dev2
frontend==0.0.3
frozendict==2.4.6
h11==0.14.0
html5lib==1.1
httplib2==0.22.0
idna==3.10
importlib_resources==6.4.5
isodate==0.6.1
itsdangerous==2.2.0
jmespath==1.0.1
looseversion==1.3.0
lxml==5.3.0
multitasking==0.0.11
networkx==3.4.2
nibabel==5.3.2
nipype==1.8.6
numpy==2.1.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
peewee==3.17.7
platformdirs==4.3.6
prov==2.0.1
pycparser==2.22
pydot==3.0.2
PyJWT==2.9.0
PyMuPDF==1.24.12
pyOpenSSL==24.2.1
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
s3transfer==0.10.3
scipy==1.14.1
selenium==4.25.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.12.3
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.41.0
tomlkit==0.13.2
traits==6.3.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
